[WIP] proposal for organization of llm collection
