+++ nemo/collections/llm/data/gpt/__init__.py
@@ -1,0 +1,7 @@
+from nemo.collections.llm.data.gpt.dolly import DollyDataModule
+from nemo.collections.llm.data.gpt.fine_tuning import FineTuningDataModule
+from nemo.collections.llm.data.gpt.mock import MockDataModule
+from nemo.collections.llm.data.gpt.pre_training import PreTrainingDataModule
+from nemo.collections.llm.data.gpt.squad import SquadDataModule
+
+__all__ = ["FineTuningDataModule", "SquadDataModule", "DollyDataModule", "MockDataModule", "PreTrainingDataModule"]
